Move battery HTML renderers into their own translation units
Every renderer defined inline in a *-HTML.h header gets emitted in
whichever TU instantiates it - in practice BATTERIES.cpp, which made the
factory object absorb ~129 KB of renderer code and broke per-driver
flash attribution (the size bot and reviewers see renderer cost charged
to BATTERIES.o, not to the driver that owns it). It also made
BATTERIES.cpp a mega-TU that rebuilds on every driver-HTML touch.

Move the method bodies to per-renderer .cpp files, following the
existing BMW-I3/BMW-IX/IONIQ-28/E-GMP idiom (21 renderers converted;
headers keep the class with declarations, trivial one-liners stay
inline). Battery headers that relied on transitive datalayer includes
through their HTML header now include what they use directly.

Measured on esp32devkit_330: BATTERIES.cpp.o 185,352 B -> 56,427 B of
flash sections; renderer cost now lands per driver (TESLA-HTML.o
27,550 B, ECMP-HTML.o 14,229 B, MEB-HTML.o 12,450 B). Firmware total
shrinks 7,280 B (1,911,536 -> 1,904,256) from better section GC. No
behavior change; test suite and -Werror env green.
