+++ Software/src/battery/CELLPOWER-HTML.cpp
@@ -1,0 +1,139 @@
+#include "CELLPOWER-HTML.h"
+#include "../datalayer/datalayer_extended.h"
+
+String CellpowerHtmlRenderer::get_status_html() {
+  String content;
+
+  static const char* falseTrue[2] = {"False", "True"};
+  content += "<h3>States:</h3>";
+  content += "<h4>Discharge: " + String(falseTrue[datalayer_extended.cellpower.system_state_discharge]) + "</h4>";
+  content += "<h4>Charge: " + String(falseTrue[datalayer_extended.cellpower.system_state_charge]) + "</h4>";
+  content +=
+      "<h4>Cellbalancing: " + String(falseTrue[datalayer_extended.cellpower.system_state_cellbalancing]) + "</h4>";
+  content +=
+      "<h4>Tricklecharging: " + String(falseTrue[datalayer_extended.cellpower.system_state_tricklecharge]) + "</h4>";
+  content += "<h4>Idle: " + String(falseTrue[datalayer_extended.cellpower.system_state_idle]) + "</h4>";
+  content +=
+      "<h4>Charge completed: " + String(falseTrue[datalayer_extended.cellpower.system_state_chargecompleted]) + "</h4>";
+  content +=
+      "<h4>Maintenance charge: " + String(falseTrue[datalayer_extended.cellpower.system_state_maintenancecharge]) +
+      "</h4>";
+  content += "<h3>IO:</h3>";
+  content +=
+      "<h4>Main positive relay: " + String(falseTrue[datalayer_extended.cellpower.IO_state_main_positive_relay]) +
+      "</h4>";
+  content +=
+      "<h4>Main negative relay: " + String(falseTrue[datalayer_extended.cellpower.IO_state_main_negative_relay]) +
+      "</h4>";
+  content += "<h4>Charge enabled: " + String(falseTrue[datalayer_extended.cellpower.IO_state_charge_enable]) + "</h4>";
+  content +=
+      "<h4>Precharge relay: " + String(falseTrue[datalayer_extended.cellpower.IO_state_precharge_relay]) + "</h4>";
+  content +=
+      "<h4>Discharge enable: " + String(falseTrue[datalayer_extended.cellpower.IO_state_discharge_enable]) + "</h4>";
+  content += "<h4>IO 6: " + String(falseTrue[datalayer_extended.cellpower.IO_state_IO_6]) + "</h4>";
+  content += "<h4>IO 7: " + String(falseTrue[datalayer_extended.cellpower.IO_state_IO_7]) + "</h4>";
+  content += "<h4>IO 8: " + String(falseTrue[datalayer_extended.cellpower.IO_state_IO_8]) + "</h4>";
+  content += "<h3>Errors:</h3>";
+  content +=
+      "<h4>Cell overvoltage: " + String(falseTrue[datalayer_extended.cellpower.error_Cell_overvoltage]) + "</h4>";
+  content +=
+      "<h4>Cell undervoltage: " + String(falseTrue[datalayer_extended.cellpower.error_Cell_undervoltage]) + "</h4>";
+  content += "<h4>Cell end of life voltage: " +
+             String(falseTrue[datalayer_extended.cellpower.error_Cell_end_of_life_voltage]) + "</h4>";
+  content += "<h4>Cell voltage misread: " + String(falseTrue[datalayer_extended.cellpower.error_Cell_voltage_misread]) +
+             "</h4>";
+  content +=
+      "<h4>Cell over temperature: " + String(falseTrue[datalayer_extended.cellpower.error_Cell_over_temperature]) +
+      "</h4>";
+  content +=
+      "<h4>Cell under temperature: " + String(falseTrue[datalayer_extended.cellpower.error_Cell_under_temperature]) +
+      "</h4>";
+  content += "<h4>Cell unmanaged: " + String(falseTrue[datalayer_extended.cellpower.error_Cell_unmanaged]) + "</h4>";
+  content += "<h4>LMU over temperature: " + String(falseTrue[datalayer_extended.cellpower.error_LMU_over_temperature]) +
+             "</h4>";
+  content +=
+      "<h4>LMU under temperature: " + String(falseTrue[datalayer_extended.cellpower.error_LMU_under_temperature]) +
+      "</h4>";
+  content += "<h4>Temp sensor open circuit: " +
+             String(falseTrue[datalayer_extended.cellpower.error_Temp_sensor_open_circuit]) + "</h4>";
+  content += "<h4>Temp sensor short circuit: " +
+             String(falseTrue[datalayer_extended.cellpower.error_Temp_sensor_short_circuit]) + "</h4>";
+  content += "<h4>SUB comm: " + String(falseTrue[datalayer_extended.cellpower.error_SUB_communication]) + "</h4>";
+  content += "<h4>LMU comm: " + String(falseTrue[datalayer_extended.cellpower.error_LMU_communication]) + "</h4>";
+  content += "<h4>Over current In: " + String(falseTrue[datalayer_extended.cellpower.error_Over_current_IN]) + "</h4>";
+  content +=
+      "<h4>Over current Out: " + String(falseTrue[datalayer_extended.cellpower.error_Over_current_OUT]) + "</h4>";
+  content += "<h4>Short circuit: " + String(falseTrue[datalayer_extended.cellpower.error_Short_circuit]) + "</h4>";
+  content += "<h4>Leak detected: " + String(falseTrue[datalayer_extended.cellpower.error_Leak_detected]) + "</h4>";
+  content +=
+      "<h4>Leak detection failed: " + String(falseTrue[datalayer_extended.cellpower.error_Leak_detection_failed]) +
+      "</h4>";
+  content += "<h4>Voltage diff: " + String(falseTrue[datalayer_extended.cellpower.error_Voltage_difference]) + "</h4>";
+  content +=
+      "<h4>BMCU supply overvoltage: " + String(falseTrue[datalayer_extended.cellpower.error_BMCU_supply_over_voltage]) +
+      "</h4>";
+  content += "<h4>BMCU supply undervoltage: " +
+             String(falseTrue[datalayer_extended.cellpower.error_BMCU_supply_under_voltage]) + "</h4>";
+  content +=
+      "<h4>Main positive contactor: " + String(falseTrue[datalayer_extended.cellpower.error_Main_positive_contactor]) +
+      "</h4>";
+  content +=
+      "<h4>Main negative contactor: " + String(falseTrue[datalayer_extended.cellpower.error_Main_negative_contactor]) +
+      "</h4>";
+  content +=
+      "<h4>Precharge contactor: " + String(falseTrue[datalayer_extended.cellpower.error_Precharge_contactor]) + "</h4>";
+  content +=
+      "<h4>Midpack contactor: " + String(falseTrue[datalayer_extended.cellpower.error_Midpack_contactor]) + "</h4>";
+  content +=
+      "<h4>Precharge timeout: " + String(falseTrue[datalayer_extended.cellpower.error_Precharge_timeout]) + "</h4>";
+  content += "<h4>EMG connector override: " +
+             String(falseTrue[datalayer_extended.cellpower.error_Emergency_connector_override]) + "</h4>";
+  content += "<h3>Warnings:</h3>";
+  content +=
+      "<h4>High cell voltage: " + String(falseTrue[datalayer_extended.cellpower.warning_High_cell_voltage]) + "</h4>";
+  content +=
+      "<h4>Low cell voltage: " + String(falseTrue[datalayer_extended.cellpower.warning_Low_cell_voltage]) + "</h4>";
+  content +=
+      "<h4>High cell temperature: " + String(falseTrue[datalayer_extended.cellpower.warning_High_cell_temperature]) +
+      "</h4>";
+  content +=
+      "<h4>Low cell temperature: " + String(falseTrue[datalayer_extended.cellpower.warning_Low_cell_temperature]) +
+      "</h4>";
+  content +=
+      "<h4>High LMU temperature: " + String(falseTrue[datalayer_extended.cellpower.warning_High_LMU_temperature]) +
+      "</h4>";
+  content += "<h4>Low LMU temperature: " + String(falseTrue[datalayer_extended.cellpower.warning_Low_LMU_temperature]) +
+             "</h4>";
+  content +=
+      "<h4>SUB comm interf: " + String(falseTrue[datalayer_extended.cellpower.warning_SUB_communication_interfered]) +
+      "</h4>";
+  content +=
+      "<h4>LMU comm interf: " + String(falseTrue[datalayer_extended.cellpower.warning_LMU_communication_interfered]) +
+      "</h4>";
+  content +=
+      "<h4>High current In: " + String(falseTrue[datalayer_extended.cellpower.warning_High_current_IN]) + "</h4>";
+  content +=
+      "<h4>High current Out: " + String(falseTrue[datalayer_extended.cellpower.warning_High_current_OUT]) + "</h4>";
+  content += "<h4>Pack resistance diff: " +
+             String(falseTrue[datalayer_extended.cellpower.warning_Pack_resistance_difference]) + "</h4>";
+  content +=
+      "<h4>High pack resistance: " + String(falseTrue[datalayer_extended.cellpower.warning_High_pack_resistance]) +
+      "</h4>";
+  content += "<h4>Cell resistance diff: " +
+             String(falseTrue[datalayer_extended.cellpower.warning_Cell_resistance_difference]) + "</h4>";
+  content +=
+      "<h4>High cell resistance: " + String(falseTrue[datalayer_extended.cellpower.warning_High_cell_resistance]) +
+      "</h4>";
+  content += "<h4>High BMCU supply voltage: " +
+             String(falseTrue[datalayer_extended.cellpower.warning_High_BMCU_supply_voltage]) + "</h4>";
+  content += "<h4>Low BMCU supply voltage: " +
+             String(falseTrue[datalayer_extended.cellpower.warning_Low_BMCU_supply_voltage]) + "</h4>";
+  content += "<h4>Low SOC: " + String(falseTrue[datalayer_extended.cellpower.warning_Low_SOC]) + "</h4>";
+  content += "<h4>Balancing required: " +
+             String(falseTrue[datalayer_extended.cellpower.warning_Balancing_required_OCV_model]) + "</h4>";
+  content +=
+      "<h4>Charger not responding: " + String(falseTrue[datalayer_extended.cellpower.warning_Charger_not_responding]) +
+      "</h4>";
+
+  return content;
+}
