+++ Software/src/battery/CMFA-EV-HTML.h
@@ -1,9 +1,9 @@
-#ifndef _TESLA_HTML_H
-#define _TESLA_HTML_H
+#ifndef _CMFA_EV_HTML_H
+#define _CMFA_EV_HTML_H
 
 #include "../devboard/webserver/BatteryHtmlRenderer.h"
 
-class TeslaHtmlRenderer : public BatteryHtmlRenderer {
+class CmfaEvHtmlRenderer : public BatteryHtmlRenderer {
  public:
   String get_status_html();
 };
