+++ Software/src/battery/GEELY-GEOMETRY-C-BATTERY.h
@@ -1,6 +1,7 @@
 #ifndef GEELY_GEOMETRY_C_BATTERY_H
 #define GEELY_GEOMETRY_C_BATTERY_H
 #include "../datalayer/datalayer.h"
+#include "../datalayer/datalayer_extended.h"
 #include "CanBattery.h"
 #include "GEELY-GEOMETRY-C-HTML.h"
