+++ Software/src/battery/MEB-BATTERY.h
@@ -1,5 +1,7 @@
 #ifndef MEB_BATTERY_H
 #define MEB_BATTERY_H
+#include "../datalayer/datalayer.h"
+#include "../datalayer/datalayer_extended.h"
 #include "CanBattery.h"
 #include "MEB-HTML.h"
